Move identity-iota diff structs into their own submodule
