+++ identity-iota/src/diff/diff_message.rs
@@ -16,7 +16,7 @@ use identity_did::verification::MethodUriType;
 use identity_did::verification::TryMethod;
 
 use crate::did::IotaDID;
-use crate::document::DiffIotaDocument;
+use crate::diff::DiffIotaDocument;
 use crate::document::IotaDocument;
 use crate::error::Result;
 use crate::tangle::MessageId;
