+++ identity-iota/src/diff/mod.rs
@@ -1,0 +1,10 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub use self::diff_iota_document::DiffIotaDocument;
+pub use self::diff_iota_document_metadata::DiffIotaDocumentMetadata;
+pub use self::diff_message::DiffMessage;
+
+mod diff_iota_document;
+mod diff_iota_document_metadata;
+mod diff_message;
