+++ identity-iota/src/document/mod.rs
@@ -1,9 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub use self::diff_iota_document::DiffIotaDocument;
-pub use self::diff_iota_document_metadata::DiffIotaDocumentMetadata;
-pub use self::diff_message::DiffMessage;
 pub use self::iota_document::IotaDocument;
 pub use self::iota_document::IotaDocumentSigner;
 pub use self::iota_document::IotaDocumentVerifier;
@@ -11,9 +8,6 @@ pub use self::iota_document_metadata::IotaDocumentMetadata;
 pub use self::iota_verification_method::IotaVerificationMethod;
 pub use self::resolved_iota_document::ResolvedIotaDocument;
 
-mod diff_iota_document;
-mod diff_iota_document_metadata;
-mod diff_message;
 mod iota_document;
 mod iota_document_metadata;
 mod iota_verification_method;
